+++ apps/web/src/components/thread/ThreadPanel.svelte
@@ -2,6 +2,7 @@
   import { onDestroy, tick } from "svelte";
   import Avatar from "../avatar/Avatar.svelte";
   import { enhanceMarkdown } from "../../lib/actions/markdown";
+  import { enhanceMentions } from "../../lib/actions/mention-highlight";
   import {
     handleLabel,
     isDeletedBot,
@@ -30,6 +31,7 @@
     reactionController: ReactionController;
     reactionsDisabled?: boolean;
     mentionPeople?: User[];
+    mentionAttentionUserID?: string;
     replyDisabled?: boolean;
     headerLabel?: string;
     headerDetail?: string;
@@ -66,6 +68,7 @@
     reactionController,
     reactionsDisabled = false,
     mentionPeople = [],
+    mentionAttentionUserID,
     replyDisabled = false,
     headerLabel = "Thread",
     headerDetail,
@@ -435,7 +438,11 @@
           onSave={() => saveEdit(root)}
         />
       {:else}
-        <div class="markdown" use:enhanceMarkdown>{@html markdown(root.body)}</div>
+        <div
+          class="markdown"
+          use:enhanceMarkdown
+          use:enhanceMentions={{ people: mentionPeople, attentionUserID: mentionAttentionUserID }}
+        >{@html markdown(root.body)}</div>
         {#if root.edited_at}
           <span class="message-edit__indicator" title="Edited {time(root.edited_at)}">(edited)</span>
         {/if}
@@ -565,7 +572,11 @@
             />
           {:else}
             <QuoteBlock message={reply} onJump={onJumpToQuote} />
-            <div class="markdown" use:enhanceMarkdown>{@html markdown(reply.body)}</div>
+            <div
+              class="markdown"
+              use:enhanceMarkdown
+              use:enhanceMentions={{ people: mentionPeople, attentionUserID: mentionAttentionUserID }}
+            >{@html markdown(reply.body)}</div>
             {#if reply.edited_at}
               <span class="message-edit__indicator" title="Edited {time(reply.edited_at)}">(edited)</span>
             {/if}